Task-40069 : when someone call deleteTeamRoom webservice, the room seems deleted
Prior to this fix, when someone which is not the room owner can call the deleteTeamRoom web service
This function doesnt delete the room, but send an event to the interface of the owner to no mmore display the room in romm list, like if the room were deleted
This fix add a check to not send event if room is not deleted.

+++ server-embedded/src/main/java/org/exoplatform/chat/server/ChatServer.java
@@ -475,7 +475,12 @@ public class ChatServer
       return Response.notFound("Petit malin !");
     }
     try {
-      chatService.deleteTeamRoom(room, user);
+      String creator = chatService.getTeamCreator(room);
+      if (!creator.equals(user)) {
+        return Response.notFound("");
+      } else {
+        chatService.deleteTeamRoom(room, user);
+      }
     } catch (Exception e) {
       LOG.log(Level.SEVERE, "Impossible to delete Team Room [" + room + "] : " + e.getMessage(), e);
       return Response.content(500, "Oups!");
